+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.0.3-pre
-*
+* Add list-dependencies, list-devdependencies, list-alldependencies commands
 
 ### 0.0.2
 * Add range version support to dependencies upgrade